+++ src/objects/Renderanle.ts
@@ -1,0 +1,5 @@
+import * as PIXI from 'pixi.js'
+
+export interface Renderable {
+  get render(): PIXI.DisplayObject
+}
